Fixed automated tests, minor updates so middle number was updated.
